Extract points_left to separate header
